move jetstream metric to new file
